Add OpenJDK as a broken static PIE application
Patch the build system so that it accepts `--enable-static-build`
and `--with-x=no`

The build is expected to fail at the linking stage. As workaround,
we run a manually edited command that creates `java`.

In this state, the generated executable fails to run even on the
host machine due to a crash inside the startup code that is
supposed to do self-relocation.

Signed-off-by: Cosmin Vancea <csvancea@gmail.com>